Nipuna

+++ src/main/java/servlet/pahanaedu/servlet/admin/DeleteUsers.java
@@ -1,5 +1,6 @@
 package servlet.pahanaedu.servlet.admin;
 
+import servlet.pahanaedu.model.User;
 import servlet.pahanaedu.service.UserService;
 
 import javax.servlet.ServletException;
@@ -7,16 +8,24 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.sql.SQLException;
 
 @WebServlet("/admin/deleteUser")
 public class DeleteUsers extends HttpServlet {
+
     private UserService userService = new UserService();
 
-
-
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+        HttpSession session = request.getSession();
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null || !"ADMIN".equalsIgnoreCase(loggedInUser.getRole())) {
+            response.sendRedirect("../login.jsp");
+            return;
+        }
         int userId = Integer.parseInt(request.getParameter("id"));
         try {
             userService.deleteUserById(userId);
